+++ src/caldate/caldate_test.go
@@ -5,11 +5,11 @@ import (
 )
 
 func Test_ConvertToSecond_Input_182_Should_Be_15724800(t *testing.T) {
-	expected := 15724800
+	expected := uint64(15724800)
 	days := 182
 	actual := convertToSecond(days)
 
 	if actual != expected {
-		t.Errorf("expected %d but got %d", expected, actual)
+		t.Errorf("expected %d but got %d", uint64(expected), uint64(actual))
 	}
 }
